fix: correct API endpoint paths and error handling
- Fix string interpolation in AuthService base URL
- Improve error message in ImageService to include HTTP status code
- Remove redundant SecureStorage class (functionality moved elsewhere)

The changes improve error handling and fix URL construction for API endpoints.
