--- lib/utils/secure_storage.dart
@@ -1,17 +1,0 @@
-import 'package:flutter_secure_storage/flutter_secure_storage.dart';
-
-class SecureStorage {
-  final _storage = const FlutterSecureStorage();
-
-  Future<void> storeToken(String token) async {
-    await _storage.write(key: 'jwt_token', value: token);
-  }
-
-  Future<String?> getToken() async {
-    return await _storage.read(key: 'jwt_token');
-  }
-
-  Future<void> deleteToken() async {
-    await _storage.delete(key: 'jwt_token');
-  }
-}
